+++ test/models/component_copywith_test.dart
@@ -257,8 +257,6 @@ Widget _overlayPortal(
   required List<KalenderEvent> events,
   required int numberOfHiddenRows,
   required double tileHeight,
-  required RenderBoxCallback getMultiDayEventLayoutRenderBox,
-  required MultiDayOverlayEventTileBuilder overlayTileBuilder,
   required OverlayBuilders? overlayBuilders,
 }) => const SizedBox();